Interpolate the box name when starting service
We'd like to be able to tag images with their commit SHA, push to a
container registry and pull down this specific image in a subsequent
pipeline and run it as a service. In order to do that, we need the `tag`
field of the service configuration to be interpolated. Any issues with
this implementation